refactor: move net spec helpers into helper file
