+++ spec/api-net-spec.ts
@@ -4,10 +4,10 @@ import { net, session, ClientRequest, BrowserWindow, ClientRequestConstructorOpt
 import * as http from 'node:http';
 import * as url from 'node:url';
 import * as path from 'node:path';
-import { Socket } from 'node:net';
-import { defer, listen } from './lib/spec-helpers';
+import { defer } from './lib/spec-helpers';
 import { once } from 'node:events';
 import { setTimeout } from 'node:timers/promises';
+import { collectStreamBody, collectStreamBodyBuffer, getResponse, respondNTimes, respondOnce } from './lib/net-helpers';
 
 // See https://github.com/nodejs/node/issues/40702.
 dns.setDefaultResultOrder('ipv4first');
@@ -29,96 +29,13 @@ function randomString (length: number) {
   return buffer.toString();
 }
 
-async function getResponse (urlRequest: Electron.ClientRequest) {
-  return new Promise<Electron.IncomingMessage>((resolve, reject) => {
-    urlRequest.on('error', reject);
-    urlRequest.on('abort', reject);
-    urlRequest.on('response', (response) => resolve(response));
-    urlRequest.end();
-  });
-}
-
-async function collectStreamBody (response: Electron.IncomingMessage | http.IncomingMessage) {
-  return (await collectStreamBodyBuffer(response)).toString();
-}
-
-function collectStreamBodyBuffer (response: Electron.IncomingMessage | http.IncomingMessage) {
-  return new Promise<Buffer>((resolve, reject) => {
-    response.on('error', reject);
-    (response as NodeJS.EventEmitter).on('aborted', reject);
-    const data: Buffer[] = [];
-    response.on('data', (chunk) => data.push(chunk));
-    response.on('end', (chunk?: Buffer) => {
-      if (chunk) data.push(chunk);
-      resolve(Buffer.concat(data));
-    });
-  });
-}
-
-async function respondNTimes (fn: http.RequestListener, n: number): Promise<string> {
-  const server = http.createServer((request, response) => {
-    fn(request, response);
-    // don't close if a redirect was returned
-    if ((response.statusCode < 300 || response.statusCode >= 399) && n <= 0) {
-      n--;
-      server.close();
-    }
-  });
-  const sockets: Socket[] = [];
-  server.on('connection', s => sockets.push(s));
-  defer(() => {
-    server.close();
-    for (const socket of sockets) {
-      socket.destroy();
-    }
-  });
-  return (await listen(server)).url;
-}
-
-function respondOnce (fn: http.RequestListener) {
-  return respondNTimes(fn, 1);
-}
-
-let routeFailure = false;
-
-respondNTimes.toRoutes = (routes: Record<string, http.RequestListener>, n: number) => {
-  return respondNTimes((request, response) => {
-    if (Object.hasOwn(routes, request.url || '')) {
-      (async () => {
-        await Promise.resolve(routes[request.url || ''](request, response));
-      })().catch((err) => {
-        routeFailure = true;
-        console.error('Route handler failed, this is probably why your test failed', err);
-        response.statusCode = 500;
-        response.end();
-      });
-    } else {
-      response.statusCode = 500;
-      response.end();
-      expect.fail(`Unexpected URL: ${request.url}`);
-    }
-  }, n);
-};
-respondOnce.toRoutes = (routes: Record<string, http.RequestListener>) => respondNTimes.toRoutes(routes, 1);
-
-respondNTimes.toURL = (url: string, fn: http.RequestListener, n: number) => {
-  return respondNTimes.toRoutes({ [url]: fn }, n);
-};
-respondOnce.toURL = (url: string, fn: http.RequestListener) => respondNTimes.toURL(url, fn, 1);
-
-respondNTimes.toSingleURL = (fn: http.RequestListener, n: number) => {
-  const requestUrl = '/requestUrl';
-  return respondNTimes.toURL(requestUrl, fn, n).then(url => `${url}${requestUrl}`);
-};
-respondOnce.toSingleURL = (fn: http.RequestListener) => respondNTimes.toSingleURL(fn, 1);
-
 describe('net module', () => {
   beforeEach(() => {
-    routeFailure = false;
+    respondNTimes.routeFailure = false;
   });
   afterEach(async function () {
     await session.defaultSession.clearCache();
-    if (routeFailure && this.test) {
+    if (respondNTimes.routeFailure && this.test) {
       if (!this.test.isFailed()) {
         throw new Error('Failing this test due an unhandled error in the respondOnce route handler, check the logs above for the actual error');
       }
